TOPIC-16: Console tag related code
